Update_220421
complete the package of garch_2
