+++ .idea/.name
@@ -1,0 +1,1 @@
+garch_2.py
