+++ pyfeng/ex.py
@@ -3,7 +3,7 @@
 # SV models (CMC, AE) from ASP 2021
 from .heston_mc import HestonMcAndersen2008, HestonMcGlassermanKim2011, HestonMcTseWan2013, HestonMcChoiKwok2023
 from .sv32_mc import Sv32McCondQE, Sv32McAe
-from .garch import GarchMcTimeStep, GarchUncorrBaroneAdesi2004
+from .garch_2 import GarchMcTimeStep, GarchUncorrBaroneAdesi2004
 
 # SABR / OUSV models for research
 from .sabr_int import SabrCondQuad
@@ -13,3 +13,6 @@ from .ousv import OusvSchobelZhu1998, OusvMcTimeStep, OusvMcChoi2023
 # Basket-Asian from ASP 2021
 from .multiasset_Ju2002 import BsmBasketAsianJu2002, BsmContinuousAsianJu2002
 from .asian import BsmAsianLinetsky2004, BsmAsianJsu
+
+# SV models: GARCH Diffusion from ASP 2022
+from .garch_2 import GarchCapriotti2018
