--- pyfeng/garchmodel/__init__.py
@@ -1,2 +1,0 @@
-from . import capriotti2018
-from . import generate_rv
